backend/pyproject.toml
[project]
name = "saq-sommelier"
version = "1.0.0"
description = "AI-powered wine recommendation platform — SAQ catalog scraper, Claude RAG, Telegram bot & web app"
requires-python = ">=3.12,<4.0"
dynamic = ["dependencies"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
fastapi = ">=0.115"
uvicorn = { version = ">=0.34", extras = ["standard"] }
sqlalchemy = "^2.0"
asyncpg = "^0.31"
# greenlet: Required by SQLAlchemy for async operations (manages coroutine context switching)
greenlet = "^3.3.2"
anthropic = ">=0.52"
loguru = ">=0.7"
saq-core = {path = "../core", develop = true, extras = ["embeddings"]}
pyjwt = "^2.11.0"

[tool.poetry.group.dev.dependencies]
httpx = ">=0.28"
pip-audit = ">=2.7"
pytest = ">=8.0"
pytest-asyncio = ">=0.25"
pytest-cov = ">=6.0"
ruff = "^0.15.5"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["backend"]
omit = ["tests/*", "eval/*"]

[tool.coverage.report]
show_missing = true
fail_under = 85

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["B", "E", "F", "I", "N", "RET", "RUF", "S", "SIM", "TCH", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["backend", "core"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B008", "S101", "S105"]
"__main__.py" = ["S104"]
"api/*" = ["B008"]
"auth.py" = ["B008"]
"schemas/*" = ["S105"]
"eval/*" = ["S105"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
